namespace-node-affinity-tester/requirements.txt
jsonpatch
requests
pyyaml
typer
